Set the digest_function field as part of all relevant gRPC requests
In the following PR we extended most of the REv2 *Request messages to
take an explicit digest function:

https://github.com/bazelbuild/remote-apis/pull/235

This eliminates the ambiguity between digest functions that have the
same hash length (e.g., MD5 and MURMUR3, SHA256 and SHA256TREE). This
change extends the REv2 client in Bazel to set the digest_function field
explicitly, so that the server does not need to use any heuristics to
pick a digest function when processing requests.

As we are now going to call DigestUtil.getDigestFunction() far more
frequently, alter DigestUtil to precompute the value upon construction.