Fix state tracking for actions that run under multiple strategies.
Each strategy can independently progress through the sequence of phases, but from a UI standpoint, we only care about (1) the set of strategies that have reported progress so far, and (2) the latest phase reported across all strategies.

In addition, each action should contribute at most one unit to the action count, regardless of the number of strategies applied to it.

This is not only simpler to track, but leads to more informative output in certain contexts. For example, if an action is simultaneously looked up in a remote cache and executed locally, we should not hide the fact that it was looked up once local execution starts (which could be an extremely short timespan). This is currently worked around by emitting the "running" event after the cache lookup is done [1], which is arguably incorrect since the action isn't running yet at that point.

Related to #7345.

[1] https://cs.opensource.google/bazel/bazel/+/master:src/main/java/com/google/devtools/build/lib/remote/RemoteSpawnCache.java;l=155;drc=32e4f23ec67909e92313be4d3334b62ccb490d55

PiperOrigin-RevId: 516782203
Change-Id: I710ac7199ec7a5fa5911788e3343445d5e57c85a