Delete downloadFile in favor of prefetchFiles in AbstractActionInputPrefetcher.
This simplifies the interface and makes it less likely that we get the subtle synchronization logic wrong in the future.

Due to these changes, AbstractActionInputPrefetcher is now too low-level to handle the presentation of a BulkTransferException to the user, so the logic to convert it into an EnvironmentalExecException is lifted into AbstractSpawnStrategy.

PiperOrigin-RevId: 516816172
Change-Id: I6e4b57fb90d9c84821dee36ae29c92cee02a906f