Don't store a parsed `RuleVisibility` in `Rule`.
When a rule is loaded, only validate the attribute value. Parse the raw attribute value into a `RuleVisibility` on demand when needed.

This saves a lot of memory from `PackageGroupsRuleVisibility` instances, many of which may belong to rules that are loaded but never depended on. This also removes a field from `Rule`. Although it doesn't move the class below an 8-byte threshold yet, I have more planned changes that will.

`bind` rules are special-cased in `Rule` since they are always public and have no `visibility` attribute. See the change in `WorkspaceFactoryHelper`, which needed to be accounted for.

In order to offset some of the cpu/garbage cost of computing `RuleVisibility` on demand, provide an alternative, more efficient method `getVisibilityDependencyLabels`, which is used by `DependencyResolver`, which only needs the dep labels and not a whole `RuleVisibility` instance.

PiperOrigin-RevId: 517199212
Change-Id: Id76adbfc87d7359675de36cdf284e8495cd0ee77