Add `--skip_incompatible_explicit_targets` option
This adds an argument to skip incompatible targets even if they were
explicitly requested on the command line. This is useful for CI to
allow it to build changed targets from rdeps queries without needing
to filter them all through a cquery to check if they are compatible.

Closes #17403

RELNOTES: Add `--skip_incompatible_explicit_targets` option