Use macro name in Java strict deps fixups
The name part of the target label embedded in JARs by the Java rules for
the purpose of generating fixup commands for strict deps violations is
now set to the macro name if the current target has been generated by a
macro. Since tools such buildozer operate on the level of parsed but
unevaluated BUILD files, they only see the macro name, not the name of
the generated target that violated strict deps.

This change is backwards compatible in the sense that fixup command that
worked before will continue to work: If the macro name is identical to
the target name, nothing has changed, if it differs, the fixup command
couldn't have worked before.