Make CompletionKeys stallable.
Also introduced a new interface StallableKeys. This allows for a more explicit targeting of which SkyKeys to hold up.

The discussion below assumes --experimental_skymeld_analysis_overlap_percentage=0 (zero overlap).

In a previous CL, we implemented the holding up of execution phase until x% of analysis phase is finished by queuing up execution phase tasks and wait for a go-ahead signal. Our definition of execution phase tasks was: anything that should be executed with threadPoolType==EXECUTION.

This caused some inefficiencies with Skymeld: the entry point of the execution phase is the CompletionFunctions, not ActionExecutionFunction. With the old definition of what to hold up, we allow Skyframe to create and store in the graph the CompletionKeys and many of its dependencies, while only holding up the actual action execution. This meant we had to carry the extra states from CompletionKeys (+deps) for nothing and for a long time.

This showed up when we observed the increased memory pressure throughout the "analysis phase" of Skymeld. [1]

PiperOrigin-RevId: 517933578
Change-Id: Ia0a71afec38fad577889c4ad6927e1216e4e3a67