[6.2.0] Include build-tools/X.Y.Z/{lib,lib64}/** in the minimal set of SDK files used by the Android integration tests. (#17826)
This is necessary to reenable remote caching/execution for these tests (and make CI runs faster!). In the absence of this change, the tests will fail because the aapt/appt2/aidl/zipalign tools can't be dynamically linked in the remote environment. See issue #8235.

We need to wait until this change makes it into a Bazel release before reenabling the tests, as they inherit the SDK files from the Bazel that runs them.

Related to #8235.

PiperOrigin-RevId: 517180365
Change-Id: I2baa66af15af85349187e75e3acd4e3cb5d84a49