Push down forceExclusiveIfLocalTestsInParallel() into SpawnStrategy
TestActionContext has this method right now, which is used to control
whether "exclusive-if-local" tests may run in parallel or not.
Unfortunately, in Bazel only the default implementation is used, which
always returns false. This causes tests to be run sequentially, even if
run remotely.

In order to implement this function properly, we need to call into the
underlying SpawnStrategy to ask it whether tests may run in parallel.
That's why we add an equally named method to that interface.

A single invocation of Bazel may use multiple instances of
SpawnStrategy, controlled through the --strategy flag. In our case we
need to consider the SpawnRunner associated with action mnemonic
"TestRunner". For that we need to pass down the SpawnStrategyRegistry.

Fixes: #17834