Avoid storing `LateBoundDefault` attribute values in `Rule`.
The change that produces the desired effect is to compare the attribute value with `Attribute#getDefaultValueUnchecked`. The unchecked method returns `LateBoundDefault`, not its default value. This allows the optimization added in https://github.com/bazelbuild/bazel/commit/1d03d53b56f90d26818186158fd31d41f67c2508 to also work for `LateBoundDefault`.

Rename `getRawAttrValue` to `getAttrIfStored` to better reflect that it does not fall back to the default value.

PiperOrigin-RevId: 518720504
Change-Id: I3141a1d731aaede8ff81b46ebc819cdb6af0fa0a