[Skymeld] Don't use SkyKeyComputeState to manage conflict checking in BuildDriverFunction.
When all analysis work in the build is finished, we clear the various intermediate states and shut down the executors for conflict checking. We don't expect any further conflict checking beyond this point, and until now have been using SkyKeyComputeState for that. This is a poor fit because SkyKeyComputeState doesn't guarantee that the same instance is returned when calling #compute on the same SkyKey and should only be used as a performance optimization. A bug may then occur: some conflict checking tasks may be sent to the executors still after they were shut down and resulted in a RejectedExecutionException.

This CL fixes that issue by using a dedicated Set in BuildDriverFunction that tracks whether a BuildDriverKey has been checked for conflicts or not. This set is reset after each build for consistency.

PiperOrigin-RevId: 519718108
Change-Id: I1cc8d883815514f3a8ba5f37f365ebabca9215d3