[6.2.0] Add native.module_{name,version} (#17893)
* Remove BzlmodRepoRuleHelper

The helper logic is only used in the BzlmodRepoRuleFunction so no need to have it in a separate place.
- Remove BzlmodRepoRuleHelper Interface & its implementations
- Refactor BzlmodRepoRuleFunction to use the helper logic
- Update tests and Build files accordingly

PiperOrigin-RevId: 486712547
Change-Id: I9a274a6a0afcc77be56bbe20e0c4d17c41c31c58

* Separate selection from bazel dependency graph

* Created new K/F/V for Bazel dependency graph named "BazelDepGraph"
* Refactored "BazelModuleResolution" to only run resolution (discovery, selection and checks) to create pruned and unpruned graph and not create any dependency value
* BazelDepGraphResolution calls BazelModuleResolution and extracts the dependency graph from it
* Updated tests

PiperOrigin-RevId: 499026445
Change-Id: Id1237f9d09015ffe8987d933431fccfcfa0c0963

* Add native.module_{name,version}

Extension authors often want to write some macro and change its behavior depending on which module is using it. For example, for rules_go, they want to look at the `go_deps` tags in a certain module and only allow access to repos generated from those tags.

We do this by introducing two new methods on `native`, callable only during the loading phase. They return the name and version of the module associated with the current repo. If the repo is from WORKSPACE, they return `None`. If the repo is generated by an extension, they return info about the module hosting the extension.

The implementation works by storing the "associated module" information in `RepositoryMappingValue`. I had attempted to store them in `BzlmodRepoRuleValue` or even `RepositoryDirectoryValue`, but those are not the right places since they might happen before we even evaluate the MODULE.bazel file (i.e. for non-registry overrides).

Fixes https://github.com/bazelbuild/bazel/issues/17652.

RELNOTES: Added `native.module_name()` and `native.module_version()` to allow BUILD macro authors to acquire information about which Bazel module the current repo is associated with.
PiperOrigin-RevId: 518849334
Change-Id: I06b4bc95b5a57de2412ee02544240b054c708165

* fix BUILD

* remove test case that was accidentally cherry-picked

---------

Co-authored-by: salma-samy <salmasamy@google.com>
Co-authored-by: kshyanashree <109167932+kshyanashree@users.noreply.github.com>