Create a `GcThrashingDetector`.
Unlike `RetainedHeapLimiter`, does not trigger manual GC. Can be configured with multiple limits, i.e. "OOM if there are X consecutive full GCs over threshold within Y seconds" for various values of X and Y.

Maintains a sliding window for each limit. This may have some waste for overlapping windows, but in practice the limits are expected to be small and there won't be many of them.

Not wired in via flags yet - will need to write a custom option converter.

PiperOrigin-RevId: 519825944
Change-Id: Ifbc4391ebe088bfca9b9841baccc5bc0de992022