Unconditionally clear the syscall cache at the end of the build.
The syscall cache is cleared each time there's memory pressure [1]. However, some entries may be left in the cache between the last clearing and the end of the build. These entries don't add much value since they are cleared at the beginning of the next build.

This CL performs a "flushing" of the syscall cache at the end of the build (independent of the memory pressure) to get rid of these entries and reduce the overall retained heap.

[1] https://github.com/bazelbuild/bazel/blob/30afc2547713907c06ad7ece65442215710154dd/src/main/java/com/google/devtools/build/lib/skyframe/HighWaterMarkLimiter.java#L63

PiperOrigin-RevId: 520626069
Change-Id: I53ea923c1868e90bfcf43950a4c685601f1197fe