Canonicalize use_extension label
Canonicalize the label by adding the current module's repo_name if the
label doesn't specify a repository name. This is necessary as
ModuleExtensionUsages are grouped by the string value of this label, but
later mapped to their Label representation. If multiple strings map to
the same Label, this would result in a crash.

Also enforce that `module()` is called first (if at all).

Closes #17920.

PiperOrigin-RevId: 520890201
Change-Id: Ice8e2feb0da591e3ba953f4a85284766ba599ebf