[Skymeld] Refer directly to the source of truth for whether we're in Skymeld mode.
Since https://github.com/bazelbuild/bazel/commit/986ef7b68d61b1573d9c2bb1200585d07ad24691, we've moved the checking of the Skymeld flags to
SkymeldModule#beforeCommand. The source of truth can afterwards be found in
CommandEnvironment.

The issue: BlazeModules are implicitly order-dependent. If we just make a copy
of the boolean value that tells us whether we're in Skymeld mode or not, that
value could be stale. A specific example here would be
BuildEventServiceModule#beforeCommand is run before SkymeldModule#beforeCommand
and carries with it the stale default value.

This CL fixes that issue by making the references to this source of truth more
direct via a Supplier.

PiperOrigin-RevId: 521692227
Change-Id: I02fbc7d5556e464ab47d08493f747b88116d9391