Add profiling of long wait times for cancelling the other branch in dynamic execution.
Long wait times are a sign that the spawn runner is doing something uninterruptible, which is a bad thing.

This (and a hack outputting stack traces of culprits) helped me find a number of places where interrupts were not handled correctly. To measure the effect, I ran a Blaze with this CL but with/without fixes of these interrupt problems, using dynamic execution for all mnemonics and a 100ms execution delay. I ran this until the build time stabilized, then extracted a histogram of these times from a profile. This is not a realistic usage, but should be a good stand-in for incremental builds, and shows the difference very well.

To find slow cancellations, I grabbed the stack trace of the other branch just before cancelling it (suspending the thread in the meanwhile to not have it move on while capturing the trace) and output it if the cancel was slow. My fixes are mainly centered on places that do a lot of work without checking interrupts.  Another and more insidious source of problems is places that swallow interrupts, I did not locate such a case yet, but there are still slow interrupts. Such cases could be distinguished by the cancelled branch actually making it to the end.

Without fixes (all times in ms):
     0 -   100: 565
   100 -   200: 877
   200 -   300: 577
   300 -   400: 252
   400 -   500: 94
   500 -   600: 138
   600 -   700: 89
   700 -   800: 101
   800 -   900: 59
   900 -  1000: 30
  1000 -  1100: 9
  1100 -  1200: 1
  1200 -  1300: 11
  1300 -  1400: 17
  1400 -  1500: 2
  1500 -  1600: 1
  1600 -  1700: 12
  1700 -  1800: 5
  1800 -  1900: 6
  2000 -  2100: 8
  2100 -  2200: 10
  2200 -  2300: 2
  2300 -  2400: 12
  2400 -  2500: 10
  2500 -  2600: 10
  2600 -  2700: 9
  2700 -  2800: 9
  2800 -  2900: 6
  2900 -  3000: 11
  3100 -  3200: 1
  3200 -  3300: 2
  3500 -  3600: 1
  3600 -  3700: 2
  3700 -  3800: 2
  4500 -  4600: 1
  4900 -  5000: 1

With fixes:
     0 -   100: 338
   100 -   200: 57
   200 -   300: 33
   300 -   400: 7
   400 -   500: 8
   500 -   600: 1
   800 -   900: 1
  1300 -  1400: 2
  2300 -  2400: 1

Tried this with a heavy build from b/266954883, got 31 extra lines on a high-performance desktop and 73 extra lines on a smaller cloudtop, so it won't grow the profile very much.

PiperOrigin-RevId: 521733420
Change-Id: I829f338ce2b8f1fb662000bb88e37d2c93a1866a