Make a rule's `Location` consistent regardless of whether `generator_*` attributes are explicitly set.
The logic that detects whether `generator_name` or `generator_function` are already set dates back to python pre-processing and is no longer relevant, but it's still being used by rules that explicitly set one or both of these attributes. Remove this check to ensure that the location is determined consistently: from the outermost frame of the call stack.

Deriving the location directly from the call stack opens up the opportunity to avoid storing the `<toplevel>` frame in `CallStack`, since it is the same as the location.

Explicitly set `generator_*` attributes seems to happen in practice when attributes are copied from `native.existing_rule`. I will look into whether this can be cleaned up or hand it off if there's a volunteer.

RELNOTES: The location of rules that explicitly specify `generator_name` and/or `generator_function` attributes (typically because they are incidentally copied from `native.existing_rule()`) is now the top-level call in the `BUILD` file, which is consistent with rules that do not explicitly specify these attributes.
PiperOrigin-RevId: 521878142
Change-Id: Ie8fb7fa9c4bef9137f6e7d7a5c50fb049024a585