Don't store `WARNING` events in Skyframe.
This means that warnings will no longer be replayed on incremental builds. A similar change was made in https://github.com/bazelbuild/bazel/commit/6e945e8ab422ca0a195d09899cce61ba6c0cfa59 for `DEBUG` events (Starlark print statements).

This is an intentional behavior change made after careful consideration and collaboration. There is a significant memory cost to store events: in addition to retaining the events themselves, each node transitively depending on an event-emitting node must have its `SkyValue` wrapped in `ValueWithMetadata`. This means that adding a single deprecation warning to a low-level target can have a surprisingly enormous memory cost.

Moreover, the use of `--auto_output_filter=subpackages` means that there is a high probability that most warnings will never even be displayed to the user even if stored and replayed.

Stored events are naturally deduplicated because they are stored in a `NestedSet`. To maintain uniqueness, additionally deduplicate warnings. Duplicate warnings may occur if a target is analyzed in multiple configurations, as shown in the test case added to `MiscAnalysisTest`.

RELNOTES: Warnings (most notably those associated with the `deprecation` rule attribute) are no longer replayed on subsequent invocations unless the target in question is re-analyzed. Warnings are purely informational, so this change has no bearing on the correctness of the build. Downstream tests that break due to this change should update their expectations.
PiperOrigin-RevId: 521879494
Change-Id: Ic4e0a68b965ed092b234fe746eeb6ea0b5562f93