python: port PyRuntimeInfo to Starlark and enable it.
Because providers are based on identity, and there are Java tests that need
to reference the provider, it's hard to split up the changes, so it's all
in a single change.

The Java class is kept to make usage by the Java tests easier; it isn't
actually used outside of tests.

Work towards #15684

PiperOrigin-RevId: 522119112
Change-Id: I3fef9b0477c110b5f32f8e26612c04bbb92ecb7b