[Skymeld] De-dup events at BuildDriverFunction.
TopLevelStatusEvents are fired from within BuildDriverFunction, that means there may be duplicated events (e.g. from resets). Up till now, we've been asking the subscribers to do their own de-duplication. This is not ideal since 1) storing the events in deduplicating sets may prevent transitive references from being GCed and 2) Posting the events were observed to be a hot spot for contention [1].

This CL moves the de-duplication into BuildDriverFunction. We exploit the fact that each BuildDriverKey only has 1 effective event per TopLevelStatusEvent type (e.g. a target foo can't be analyzed twice in a build) to prevent sending a 2nd event of the same type. This allows us to store only the type of the sent events, instead of the events themselves.

PiperOrigin-RevId: 522334018
Change-Id: I76aed9fa68f8197d8f0a70158198c7341dc3a884