Track a single `boolean containsErrors` at the package level instead of for each rule.
There is already a policy that all rules in a package are considered to be in error if the package is in error, so it is not necessary to store a `boolean` for each rule.

Removing the field from `Rule` doesn't move it below an 8-byte threshold, but it does open up enough room to add a 4-byte field for free. I am considering using this to store the `generator_name`, which is currently stored as an attribute value for every rule.

PiperOrigin-RevId: 522374259
Change-Id: I55cdaf6d34e8517145de125fa664796423d914af