Greatly reduce the memory cost of storing rule instantiation call stacks.
Memory savings are from two optimizations:

1. Remove the `CallStack` wrapper in favor of referencing the `Node` directly. The wrapper stored the size of the stack, but the size was only ever used to distinguish between singleton and multi-element call stacks. This can be done in constant time without tracking the size. This saves 24 bytes per rule.
2. The first `Node` in the call stack is redundant with the Rule's `Location`. Only store the *interior* nodes. This saves 32 bytes per rule.

This makes the cost of storing rule instantiation call stacks pretty negligible, since the interior nodes have a high rate of reuse.

An alternative considered was to continue storing the outermost call as a `Node` and reconstruct the `Location` on-demand. Although this would eliminate a field from `Rule`, this is a net loss vs the chosen approach because:

* `Rule#getLocation` is called when the rule references input files, so a new `Location` object would get created for each `InputFile`.
* There is virtually no reuse among outermost nodes, since the whole call stack must be identical. On the other hand, `Location` can be reused since it has no `next` pointer involved in equality. It's also reused naturally because it's not popped from the call stack during Starlark evaluation.

One important change is that rule serialization must preserve the rule's `Location` since it is not reflected in the compact call stack. The `Location` is appended at the head of the interior call stack at time of serialization so that it can take advantage of the string memoization table. Upon deserialization, the head is extracted and used for the rule's `Location`, leaving the tail for the interior.

PiperOrigin-RevId: 522463593
Change-Id: Ia5e8815cdff7f2838295456b4c57276b21df4d93