Check target incompatibility before toolchain resolution
Known missing toolchains are a valid reason for declaring a target incompatible and should not result in the target failing to build rather than being skipped as incompatible.

This is realized by moving the check for target incompatibility to the beginning of `PrerequisiteProducer#evaluate`, which requires fetching the target platform and config conditions.

Closes #17947.

PiperOrigin-RevId: 522534102
Change-Id: I9870edbf9cdb9873641f63610c86eb642f5c1d29