python: port PyCcLinkParamsProvider to Starlark and enable it.
Because providers are based on identity, and there are Java tests that need
to reference the provider, it's hard to split up the changes, so it's all
in a single change.

The Java class is kept to make usage by the Java tests easier; it isn't
actually used outside of tests.

Work towards #15897

PiperOrigin-RevId: 523182610
Change-Id: I6e07cf55ba08bb87c5ccb6cd37b473ed6aeeb760