Make `getPrimaryOutput()` always return the first artifact in `getOutputs()`.
This is already the case everywhere except `CppCompileAction` but was not documented, leading to `SpawnAction` unnecessarily storing a field for the primary output when it's already just the first element in its outputs. This change saves 8 bytes per `SpawnAction` and `GenRuleAction`, and moves other `SpawnAction` subclasses closer to an 8-byte threshold.

`CppCompileAction` had been using the coverage artifact (if present) as the first output but not the primary output. This is easy to change by storing the coverage artifact in a field instead of the primary output. It also removes a boolean field since we can just check the nullness of the coverage artifact field, although this does not change the shallow heap of `CppCompileAction`.

PiperOrigin-RevId: 523196083
Change-Id: Id6fe4ef126334a2b4cd2d2dc63409e6668e5c8bd