Split MetadataProvider.getMetadata() into two methods.
One for getting metadata for the inputs of the action and one for its outputs.

The principal difference between the two is that when the action starts
executing, the metadata for its inputs should be known, so no expensive I/O
should be needed. This would in principle allow one to not throw IOException
there. It remains for two reasons:

* Some conditions are signaled by throwing FileNotFoundException (I haven't
  investigated very deeply)
* Some implementations still do I/O for reasons unknown

At the very least, this will allow one to not throw InterruptedException in
getInputMetadata().

Work towards fixing #17009.

RELNOTES: None.
PiperOrigin-RevId: 523318267
Change-Id: Ib3af4c099a0faafb9a8b6d75d04928af47bfbcd1