Make MetadataHandler not implement MetadataProvider anymore.
Before this change, ActionExecutionContext got two references to a
MetadataProvider implementation: one directly, one behind MetadataHandler.
These could and did get out sync after e.g. calls to withOutputsAsInputs().

This change fixes that. The game plan is to rename these two so that it's
obvious that one deals with the inputs of the action, the other, with the
outputs, but let's do that very mechanical change after this slightly less
mechanical one.

RELNOTES: None.
PiperOrigin-RevId: 523408132
Change-Id: I557cc5ec939b7a49e33bb857c3b0916d23538048