Report failures to add permissions with JavaIoFileSystem
Previously the return value of the `File#setX` functions was not
checked, thus masking potential errors. Since these functions are
expected to fail on Windows if asked to remove a permission, errors are
only reported when attempting to add one.

Also removes misleading `throws` declarations that may mask similar
issues of this type.

Closes #18004.

PiperOrigin-RevId: 523454138
Change-Id: Ia85db5e912d07af1ca6f7cd4b8e2d784395e7811