starlark_doc_extract: stringify label-valued default attribute values
For human-readable docs, want to stringify labels in a way that
(1) elides the name part when unambiguous, matching buildifier behavior; and
(2) respects repo mapping, so we don't print full bzlmod canonical repo names.

Unfortunately, this means introducing yet another stringification method to
Label, which combines the behavior of getDisplayForm() and toShorthandString()
but cannot be conveniently composed from them. Alas.

PiperOrigin-RevId: 523504039
Change-Id: Ia09ebb77cff09aaeb270f43ed7e319c90a44aa52