python: port PyWrapCcInfo to Starlark and enable it
Because providers are based on identity, and there are Java tests that need
to reference the provider, it's hard to split up the changes, so it's all
in a single change.

The Java class is kept to make usage by the Java tests easier; it isn't
actually used outside of tests.

Work towards #15897

PiperOrigin-RevId: 523748496
Change-Id: I1f5e9316bb3c4aed6c36fbd317914e4deea26f61