Add exec_kind to RunnerCount in BEP
Solves #17296 .

The execution kind of runners is now stored in the RunnerCount message of the Build Event Protocol. For runners where execution kind is not applicable, e.g. "total", no value is set. The exec_kind field has been set as a string field rather than using an enum to avoid implicit dependencies between the Java code and the protobuf.

Closes #17585.

PiperOrigin-RevId: 524281812
Change-Id: I55518886f976397db58e71b0b34b32cf287e3d51