Add `alias_label` struct field to all targets
This allows rules to obtain the label of an `alias` target rather than
only the label of the target an alias chain ultimately resolves to. The
former is the label as specified by the user on the rule, which can be
important to know for e.g. fixup messages and manual implementations of
location expansion.