profile: add profile_finish_ts
"date" field in Bazel json profile is a confusing field.

- It uses Java Date.toString() output which includes local timezone
  that could be hard to parse.

- The name is misleading since users could mistook it for profile
  start time, while it actually means the finish/end time, when
  Bazel calls the writer to serialize profiling data to disk.

Add "profile_finish_ts" which has a clearer name and more consistent
value with the time unit being used in the rest of the JSON
profile(microseconds).

We shall deprecate the "date" field in a separate patch in a major
release.