docs: remove external links from build encyclopedia nav
These are a mix of current and stale. For example, 'JavaScript (Closure)' is used by almost no one, as far as I know.
I could imagine trying to update this instead of remove it, however there's not a good process for the Bazel team to keep updating an external 'catalog' of rules.

Note, there's a file next to the one I've edited, be-nav.vm, but I cannot find where that appears on the site. Maybe it's dead code?