Wire up `PathMapper` in `SpawnInputsExpander`
`PathMapper`s rewrite paths in command lines to make them more cache
friendly, which requires executor support to stage files at the
rewritten paths. This commit wires up the `PathMapper` used by a given
`Spawn` in `SpawnInputsExpander`, which takes care of this for inputs
to `Spawn`s executed in a sandbox or remotely.

Constructs specific to Blaze, filesets and archived tree artifacts, are
not covered by this change.

Work towards #6526