Add root path string to thinlto index files
And update the tests.

Currently when generating .thinlto.bc files and .imports files, for an object file at

[root1]/foo/bar/file

and the final executable is from

[root2]/a/b/obj

the files are stored at

[root2]/a/b/obj.lto/foo/bar/file.thinlto.bc
[root2]/a/b/obj.lto/foo/bar/file.imports

This works when [root1] and [root2] are the same; however, if they are not, it is ambiguous and we cannot know what was the origin of the index files.

With this change, now the files are stored at:

[root2]/a/b/obj.lto/[root1]/foo/bar/file.thinlto.bc
[root2]/a/b/obj.lto/[root1]/foo/bar/file.imports

PiperOrigin-RevId: 527591880
Change-Id: I30d1734f577980f1c91509d5d44858c252d3fade