Optimize `PackageGroupContents` for both memory and CPU.
The main memory improvement is to not retain `PackageSpecification` instances. The main CPU improvement is to optimize for large lists of negative package specifications, since they actually exist in practice (the old code was intentionally inefficient here).

This saves 1.5% CPU and 0.3% memory on one benchmark.

I think we can get minor code readability and GC churn improvements by ditching the `PackageSpecification` abstraction completely. I added TODOs for this. We can get further readability improvements by getting rid of the need for the `includeDoubleSlash` and `WithoutRepository` business; simplifying the requirements would allow the code to be simpler :) There are existing TODOs for part of that.

PiperOrigin-RevId: 527722770
Change-Id: Ib8efaa8b7fb0f090670a3a4e2815769d675f2669