Cpu timec to bep
This PR adds support to BEP for collecting aggregated CPU time of spawned subprocesss, grouped by mnemonic.

Main use case is to identify which mnemonics are bottlenecks and should be prioritized for optimization. Such optimizations can be done by improving associated tool chains or improving cache hit ratio for those mnemonics. Collecting the data via BEP, from all user and all CI builds, ensure that the metrics are representative in general (not only cherry-picked specific scenarios). Trends can be visualized on dashboards grouped by mnemonic.

When observing increased average builds time trends (wall clock time) for an organization, it is often not obvious to identify the cause. CPU time metrics can then not only help identifying regressions for a particular mnemonic, but by correlating CPU time with wall clock time, it can also indicate if the cause is an environmental issue or not.

Closes #15640.

Change-Id: Ic6a0f1ccea6fec9aebebc113c3d2de9e78f6b820
PiperOrigin-RevId: 527781430