Populate the ActionResult.output_symlinks field when creating the UploadManifest.
In d3fb047, we started populating the Command.output_paths field in addition to the (now deprecated, but required for backwards compatibility) output_files and output_directories fields. Accordingly, we should populate the ActionResult.output_symlinks field in addition to the deprecated output_file_symlinks and output_directory_symlinks fields.

PiperOrigin-RevId: 527869985
Change-Id: Ib1f6779299de1323bcb9c56d6ed50aa894891e4e