Amend BwoB test for action with a symlink output.
"Running a remote action with a symlink output" could mean one of two things:

(1) The action declares an output symlink which is reported as a symlink by
    the remote executor.

(2) The action declares an output non-symlink which is reported as a symlink
    by the remote executor.

While I believe (1) should be supported, I'm much less convinced about (2),
because it amounts to asking Bazel to resolve the symlink given incomplete
knowledge about the filesystem where it was produced; this is generally
impossible and potentially non-hermetic. (While we could support only hermetic
cases, e.g. by requiring it to resolve against one of the other action outputs,
that still requires a degree of sophistication that doesn't exist today.)

Therefore, this PR replaces (2) with (1) in the test added by ca30372, so
that it accurately reflects the work to be done.