Remove `CommandLineLimits` field from `SpawnAction`.
In most cases, it comes from the `BuildConfigurationValue`, which after recent changes is now accessible via the `ActionOwner`. In some cases, it is overridden to be `CommandLineLimits.UNLIMITED`, in which case subclasses can override `getCommandLineLimits()`.

`SpawnActionTemplate` now uses a subclass since it uses `UNLIMITED`. `SpawnActionTest` is updated to use an `ActionOwner` with an accurate `BuildConfigurationValue`.

PiperOrigin-RevId: 527883149
Change-Id: I6406941540155f2ea1cb3af072fec42e92828280