Preserve Starlark field declaration order in StarlarkProvider.getSchema (former getSchemaWithDocumentation)
We want the original field declaration order for starlark_doc_extract, both
to match existing Stardoc behavior and to match existing logic of preserving
the original declaration order of rule attributes etc. in proto output (leaving
sorting up to the renderer).

We want to take this opportunity to clean up StarlarkProvider's field and
method names a little bit:

* rename StarlarkProvider.schema to `fields` to emphasize that these are just
  the fields names, not the full schema given to the provider() call - it has
  a different iteration order and is missing doc strings - and to match the
  getFields() method name, logically enough
* rename StarlarkProvider.getSchemaWithDocumentation to just getSchema - since
  it now returns the original schema from Starlark with the same order, and
  since the method may be used regardless of whether the schema is documented

PiperOrigin-RevId: 528545413
Change-Id: I2756f4520c80acaca17dcd244b61d60aac5539b4