Allow overrides in non-root modules
They're simply ignored.

RELNOTES: Overrides specified by non-root modules no longer cause an error, and are silently ignored instead. They were originally treated as an error to allow for the future possibility of overrides in the transitive dependency graph working together; but we've deemed that infeasible (and even if it was, it'd be so complicated and confusing to users that it would not be a good addition).
PiperOrigin-RevId: 529095596
Change-Id: I8b9b7b570b405ee757554accf791d8e4c1ff6528
(cherry picked from commit 78cb7d5b652ee155a0d1ad5cef3a3131e9705152)
Signed-off-by: Brentley Jones <github@brentleyjones.com>