Fix crash with heuristic node dropping and partial reevaluation
InMemoryNodeEntry#isReadyToEvaluate had an assertion that crashes when
the dirtyBuildingState field is unset. This state can be reached when a
node is heuristically dropped, then recreated via getOrRecreateDepEntry,
and then observed prior to `getOrRecreateDepEntry`'s call site then
calling `depEntry.addReverseDepAndCheckIfDone`, which is responsible for
initializing the newly created NodeEntry's DirtyBuildingState. The
opportunity to observe this state happens only during a done PE
node's trailing `#run()` call, which occurs when a completed dep's
signal, which enqueues a PE node's evaluation, races with the completion
of that PE node.

The isReadyToEvaluate is productively called at only one site, and in
that context, its assertions are unnecessary.

PiperOrigin-RevId: 529425479
Change-Id: Ibf7d74fc1e295e7bbfd5dd6b1a1b2009d36c5d4b