Use canonical names to resolve relative labels in Stardoc
Previously, when a `.bzl` file was loaded via an apparent name and
triggered a load of another file via a relative label, the new file
would be loaded with that apparent name, which is incorrect.

This is solved by converting all apparent labels into canonical labels
and only then turning those into runfiles paths.

Also ensures that line endings are consistent on all platforms, which
simplifies the setup of a cross-platform test.