TransitiveActionLookupKeysCollector cleanup.
Replaces ActionLookupKeyOrProxy with ActionLookupKey in the collector.
ActionLookupKeyOrProxy is an opaque logical representation of the
key. Since the TransitiveActionLookupKeysCollector is a physical traversal
of the graph ActionLookupKey makes more sense there.

Improves the performance of transitive traversal by making it
intrusive. Doing so avoids creating redundant subtasks and additional
WalkableGraph lookups.

PiperOrigin-RevId: 529739700
Change-Id: I35f79ff4a5fbd4cb19baeca6409ddbc2bdff04eb