Don't retain `CommandLineAndParamFileInfo` wrappers in `CommandLines`.
It's cheaper to store the `CommandLine` and corresponding `ParamFileInfo` at adjacent array slots and we don't lose any type information since we're already using an `Object[]`.

Break up `CommandLines` into an abstract class to better represent the distinct cases of a single `CommandLine` and more than one. This also helps with my planned next step of optimizing for the common cases of 2-3 part command lines and command lines made up of a single type (e.g. `StarlarkCustomCommandLine`).

After taking some time to study this, I think it would actually make more sense for `CommandLines` to be called `CommandLine` and `CommandLine` to be called `CommandLinePart`. I will consider renaming in a followup.

PiperOrigin-RevId: 529865102
Change-Id: Ia6bdbc0ec3054cd503cc84520d7bf588e466f205