[Skymeld] Fix a grep in ui_test.sh
ui_test.sh greps, in an infinite while loop, the string "[2 / 3]" in the stderr
before moving on with the test. This condition won't be satisfied with Skymeld
because the workspace status actions are pushed to the back (one would see
"[1 / 2]" with Skymeld). That's the reason why the loop kept on going and
the test eventually timed out.

PiperOrigin-RevId: 530272949
Change-Id: Ica9f6579d2c42e7b82ad7f2d8f19d1d32bd5b5ef