[Skymeld] Move the workspace status actions earlier.
These actions are supposed to be run at the beginning of a build.

https://github.com/bazelbuild/bazel/commit/41f7f8aa9a7cc714fe5f12a346b91e04ec9d7561 moved these actions to the end of the build to avoid a race
condition with the CriticalPathComputer. This was however insufficient and
the race condition was finally fixed in https://github.com/bazelbuild/bazel/commit/7e9e5035190bc9ea3d672aef553dc7adf4e7c0f5, which also made it
unnecessary to delay these actions until the end. By the time we start with
evaluating the BuildDriverKeys + WorkspaceStatusArtifact, the CriticalPathComputer
is ready.

PiperOrigin-RevId: 530309785
Change-Id: Ic83e9bc5becca521f90029188bf50b1d1bd120e7