[Skymeld] Fix BazelBuildEventServiceModuleTest.
The test was failing because of a mismatch in configuration. The right one, in
the context of this test, is determined from the list of analyzed target. With
Skymeld, the first action execution may happen before any analysis is done (via
the workspace status action). If we OOM at the first action execution, then
it's likely that there's no analysis info available with Skymeld.

This CL fixes that by only OOM-ing at the first shareable action, which makes it
certain that some analysis should have already been done.

PiperOrigin-RevId: 530878137
Change-Id: Ic85e0f1f4d9b8e74abe1701869b136f6767d978d