Simplify EnvironmentExtension
This reduces the API of PackageFactory.EnvironmentExtension with an eye toward eliminating it by merging it into Bootstrap. This will help us get PackageFactory out of the business of managing the Starlark environment.

The extension's `updateWorkspace` method is removed, along with the workspace logic that called it. This feature has been unused for a long time.

The extension's `updateNative` method is also removed and merged into `update`. There are no call sites that need to differentially add a symbol to the BUILD environment but not the `native` object, or vice versa.

PackageFactory no longer needs to retain the list of extensions, since it was only retrieved for workspace logic.

PiperOrigin-RevId: 530917040
Change-Id: Ie0ff2be22e8f6c6c91a2c833dc0b50dd19b1ae85