Move EnvironmentExtension registration to ConfiguredRuleClassProvider
The ConfiguredRuleClassProvider and its builder get a new API for registering and retrieving EnvironmentExtension. This API is used by BazelStarlarkEnvironment, which is what actually determines the final environment. The old pathway going through PackageFactory and its builder-for-tests is deprecated; its removal will come in the follow-up CL to keep the diffs more readable.

This will remove a lot of boilerplate from our test initialization.

PiperOrigin-RevId: 530933441
Change-Id: I4d741d56055df76a91d44db4692a105fe69ff9c4