[Skymeld] Fix MetricsCollectorTest, for real this time.
`getCommandEnvironment().withMergedAnalysisAndExecutionSourceOfTruth()` only
has scope of what happens in a single build, whereas the variable `newGraphSize`
is influenced by all builds from a server. This led to a bug when `--nobuild`
was involved: because the last build wasn't done with Skymeld, we failed to
update `newGraphSize` when it should have been updated.

This CL fixes that by recording once whether Skymeld was involved in this series
of builds.

PiperOrigin-RevId: 531138812
Change-Id: I94ac47253a5656508a745f652cf52e8b6f65379a