Properly account for symlinks and root symlinks in repo mapping manifest
A runfiles symlink entry `path/to/pkg/file -> path/to/artifact` is
implicitly prefixed with the workspace name and should thus result in
repo mappings to the main repository to be added to the repo mapping
manifest, not mappings for the owner of the artifact.

A runfiles root symlink entry `first_segment/path/to/pkg/file ->
path/to/artifact` should result result in repo mappings to the repository
with canonical name equal to the first segment to be added to the repo
mapping manifest, if it exists.

Along the way clarifies in the documentation of
`ctx.runfiles(symlinks = ...)` that the workspace name will be prefixed
implicitly.