Remove `inputsForInvalidation` field from `CppCompileAction`.
`inputsForInvalidation` comes directly from `CcComplilationContext` which is already stored.

Additionally, change a test has to instead actually modify the CC toolchain config instead of overridding builtin include directories using a test-only builder method. An earlier version of this change aimed to store `CcToolchainProvider` in `CppCompileAction` instead of `grepIncludes` and `builtinIncludeDirectories`, but this wasn't a good idea since `CcToolchainProvider` otherwise does not survive analysis cache discard. That was reverted, but the test improvement is kept.

PiperOrigin-RevId: 531534608
Change-Id: I3df47d676f9e2cc2a8cd8ff4a5fd052e3f0c1ad4