Fix and optimize `Runfiles#fingerprint`
The fingerprint did not include the conflict policy and some of the
collection's sizes. It also didn't use the cache for fingerprints of
`NestedSet`s and instead always flattened the sets.

The new `fingerprint` method on `EmptyFilesSupplier` makes it possible
to drop the call to `Runfiles#getEmptyFilenames`, which would still end
up flattening the sets.