python: Add repo mapping to zip builds so runfiles lookup works correctly
To make runfiles lookups work correctly with bzlmod, a `_repo_mapping` file
is used to map apparent names to canonical names. This file is normally
automatically created by higher level code and injected into the runfiles, but
that happens after a rule has finished running. This means it isn't in the
runfiles object that py_binary itself sees when its constructing a zip file of
itself.

To fix, the zip file generates a repo mapping itself and puts it into the zip
file, thus allowing the runfiles library to find it when run from the zip file.
This requires a Java hook since the underlying `RepoMappingManifestAction`
isn't directly exposed to Starlark.

Fixes #17941

PiperOrigin-RevId: 532265478
Change-Id: I5f36fe58f043611481a7ed6ba19a7bbf5be4edf2