[Skymeld] Make skymeld include alias test.
Given 2 targets `//a:a` and `//a:a_alias`, we only know that one is an alias of
the other after the analysis phase (i.e. when we have the `ConfiguredTarget`
objects. In regular blaze, this information is then used to determine the test
type (parallel, exclusive, ...).

With Skymeld, we used to determine the test type before the analysis/execution
phase. This would result in `//a:a_alias` not being classified as a test.

This CL fixes that by moving the test classification to within
`BuildDriverFunction`, after analysis is done for a target.

PiperOrigin-RevId: 532769383
Change-Id: I1a4eb381fdf6abec2918a1e0e4e9f457ecb31470