Add ability to register BUILD toplevels in ConfiguredRuleClassProvider
This will take the place of EnvironmentExtension#update.

I previously considered allowing the CRCP to register BUILD toplevels that do not get added to the `native` module. But there doesn't seem to be a need for this, and it seems dangerous to allow it in absence of a compelling reason.

PiperOrigin-RevId: 532848716
Change-Id: Ia9b6d40827748cd6d3f5c4ef33737c1f4cd21b59