Fix Xcode 14.3 compatibility
With Xcode 14.3+ on x86_64 machines there is an opaque error when
linking binaries on macOS because Apple removed an old support library.
That library is only linked if the macOS target is < 10.11, so this
bumps the default versions past that. This macOS version was released in
September 2015.

Fixes https://github.com/bazelbuild/bazel/issues/18278