Improve error message if `select` references non-existent package
If a key in a `select` references a package that does not exist, the error message now includes a reference to the target with the `select`:

```
ERROR: no such package 'hello/c': BUILD file not found in any of the following directories. Add a BUILD file to a directory to mark it as a package.
 - /tmp/bazel-crash/hello/c
ERROR: /tmp/bazel-crash/hello/b/BUILD:1:13: errors encountered resolving select() keys for //hello/b:b
ERROR: Analysis of target '//hello/a:a' failed; build aborted
```

Previously, it didn't:
```
ERROR: Analysis of target '//hello/a:a' failed; build aborted: no such package 'hello/c': BUILD file not found in any of the following directories. Add a BUILD file to a directory to mark it as a package.
 - /tmp/bazel-crash/hello/c
```

Fixes #17689

Closes #18454.

PiperOrigin-RevId: 533635167
Change-Id: I714b1c1c539d9dfb76dea9b6f1899bb98ae558dd