Unify the logic to determine the set of modified output files.
Right now a `ModifiedFileSet` instance is created in `ExecutionTool#executeBuild`
and later on a piece of logic in `ExecutionTool#createBuilder` determines the
effective modified file set we'll use in this build. This is nothing wrong per
se, but resulted in a subtle bug in Skymeld where the intermediate version of
`ModifiedFileSet` was used in a subsequent part of the build, which diverges
from the effective version in `SkyframeBuilder` under specific conditions.

To avoid similar mistakes in the future, this CL groups this logic together into
a method in `ExecutionTool`.

PiperOrigin-RevId: 533988768
Change-Id: I2db8ed29ed61ba205394c7026f12c10ae5f8bf7d