Remove output group for debug files from cc_shared_library
It is not necessary to have this mechanism in the cc_shared_library rule
itself. The same data can be obtained by using aspects as shown in
https://github.com/oquenchil/bazel-contrib/tree/main/cc/tools/cc_shared_library
by the static_linker_inputs.bzl and claimed_exports.bzl files.

Removing this code from the cc_shared_library rule simplifies an already
complicated rule and potentially makes the rule more efficient since the
creation of the action for the debug files would use up memory and cpu even
when the debug files weren't actually used. This could be mitigated by a C++
configuration flag but then the debug files couldn't be added to a filegroup if
the flag was disabled.

I couldn't find anyone relying on that output group from within their build.
The Tensorflow codebase doesn't have references to it anywhere.

If you want this functionality please refer to the aspects mentioned above.
Please also keep in mind that
https://github.com/oquenchil/bazel-contrib/tree/main/cc/tools/cc_shared_library
is a personal repository and won't be actively maintained by the Bazel team.
Anything there that is generally useful can be potentially moved to
https://github.com/bazel-contrib and the actual users can help with
maintenance. The Bazel team for now can only commit to the actual core
functionality provided by the cc_* rules (including cc_shared_library) but not
to any auxiliary tools/utilities that are seldom used.

RELNOTES:none
PiperOrigin-RevId: 533990119
Change-Id: I5006113cef845d808f93f183188a4f5022a9c287