Intern lists of Javac options instead of the individual strings.
The current strategy of interning the individual strings when creating a `JavaCompileAction` is actually counterproductive (worse than no interning at all) because it takes an `ImmutableList` instance that would have been shared among multiple actions and creates a new `ImmutableList` for each action.

Instead, intern the entire list as soon as it's known to be in its final state. This effectively achieves virtually all of the benefit of interning the individual strings, plus much more.

PiperOrigin-RevId: 534066168
Change-Id: Ie61456a41fedc9e8673a8722bf49d42a2dafc865