Fix Xcode 14.3 compatibility (#18490)
With Xcode 14.3+ on x86_64 machines there is an opaque error when linking binaries on macOS because Apple removed an old support library. That library is only linked if the macOS target is < 10.11, so this bumps the default versions past that. This macOS version was released in September 2015.

Fixes https://github.com/bazelbuild/bazel/issues/18278

Closes #18460.

PiperOrigin-RevId: 534743568
Change-Id: I131880096c941df0097fe3b1faabd5a6afada4f3

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>