Utilize `CustomCommandLine#addObject` to add a `Collection` without copying over its elements.
This is more memory efficient if the `Collection` is retained anyway, since we spend a single array slot on a pointer instead of an array slot for each element. It works straightaway because `CustomCommandLine` already has handling for lazily "expanding" `Iterable`.

This change only affects Java compilation actions, as they are the only case I found which is currently wastefully copying lists. To maximize the benefit, structure the `JavaHeaderCompileAction` builder to accept an existing `ImmutableList<String>` of Javac options instead of creating its own.

PiperOrigin-RevId: 535304666
Change-Id: Id41af250916e95e87fd69bcdc1bfda1d3bc1421f