Collect --sandbox_debug output separately from stderr.
This avoids polluting the disk/remote cache entries with debug output, causing it to be replayed in subsequent cached builds.

Only non-worker Linux sandboxes are affected. Worker sandboxes, despite currently passing the debug flag to the sandbox process, never capture the worker stderr, so after this CL they will stop emitting debug output entirely. The process wrapper sandbox doesn't currently pass the debug flag, and remains that way. Finally, the Windows sandbox is experimental and provided by an external binary we don't control, so I'd rather not touch it in this CL.

Fixes #16689.

PiperOrigin-RevId: 537861949
Change-Id: Iacd8369d1c80b0dcceaad8d901d222c6592f98ce