Fix blatant bug introduced by https://github.com/bazelbuild/bazel/commit/daf80769da862e3a5f9de51f3844cbdf8f9dea66.
That CL tried to introduce a fast path in `AbstractBlazeQueryEnvironment#transformAsync` for the situation where the input `QueryTaskFuture` is already done, but it handled only the obvious sub-case where the `QueryTaskFuture` was done and successful. It's also possible for the `QueryTaskFuture` to be done and cancelled (corresponding to an interrupt being observed during evaluation of a sub-expression) or done and unsuccessful (corresponding to a failure during evaluation of a sub-expression).

PiperOrigin-RevId: 537919715
Change-Id: I8b427c494e0f8df4a206749ec1801b606a310c6a