Remove superfluous example
There is already a cpp example in bazelbuild/examples.
See: https://github.com/bazelbuild/examples/tree/main/cpp-tutorial

The one in bazelbuild/examples is referred in https://bazel.build/start/cpp
and it contains everything that this example contains.
The only part missing in the example in bazelbuild/examples is to have
a cc_test but this is covered in https://bazel.build/tutorials/cpp-use-cases#run-c-tests

This would bring us closer to https://github.com/bazelbuild/examples/issues/201