Defers TargetConfiguredEvent until after analysis.
After deleting the top-level rule transition, the configuration IDs of top
level targets will not be known until late in the analysis phase. The
TargetConfiguredEvent is now emitted after those top level targets are
analyzed.

When an analysis error occurs, may abort the TargetConfiguredEvent instead
of the yet to be announced TargetCompleteEvent.

Adds more library methods for computing BuildEvent messages for
BuildConfigurationKeys and BuildConfigurationValue. Note that
since ConfigurationId is solely based on the options checksums, the
BuildConfigurationValue isn't needed so the configurationLookupSupplier
can be deleted.

RELNOTES: (BEP) TargetConfigured events will be marked aborted instead
of published when there is an analysis error. This is motivated by a
low level Blaze change aimed at improving scalability.
PiperOrigin-RevId: 538665297
Change-Id: I191b4e26081ddbe6fc623f1f809a3045ef760a33