Emit consistent names for providers in starlark_doc_extract
If a module re-exports a provider in a namespace or under a different name, and
that provider is also referred to by a rule or rule attribute, we should use the
exported provider name consistently throughout our doc output (note that we also
output the OriginKey message with the name under which the provider was exported
in its original module).

This requires processing a module's docs in two phases: first, collect all
documentable providers' names and keys; and then process all documentable symbol
documentation as before, using the providers' exported names when emitting rule
docs.

In turn, that means refactoring ModuleInfoExtractor to use a common abstraction
(GlobalsVisitor) for recursively walking namespaced Starlark values in both phases.

Also remove the rewriteWorkspace call in originKeyFileAndModuleInfoFileLabels
which was accidentally added in the previous change.

PiperOrigin-RevId: 538939476
Change-Id: I33565966019673b58caa134e676900ba16d09412