experimental flag to disable topo sort
This is a short-term workaround, as this toplogical sort in cc_shared_library is incompatible with Android builds. This buys us some time to discuss and devise a long-term solution.

Manually tested to allow Android to build. (Along with an update for Android macros to set this new flag value to "True")

RELNOTES: None.
PiperOrigin-RevId: 539084351
Change-Id: I5a2b552f9cd158b2c45b94a2a5ba4c72c9673c2c